+++ App.js
@@ -1,10 +1,12 @@
 import React from 'react';
 
+
 // screens
-import Login from './screens/Login';
-// import Vehicles from './screens/Vehicles';
+import Login from './screens/login/Login';
+import Vehicles from './screens/vehicles/Vehicles';
 
 export default function App() {
   return <Login />;
+     <Vehicles />;
 }
